multinomialFair/requirements.txt
anytree==2.8.0
cycler==0.11.0
Cython==0.29.28
data==0.4
decorator==5.1.1
funcsigs==1.0.2
future==0.18.2
kiwisolver==1.3.1
latex==0.7.0
lxml==4.9.0
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
Pillow==8.4.0
POT==0.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.1
scipy==1.5.4
shutilwhich==1.1.0
six==1.16.0
tempdir==0.7.1
xlrd==2.0.1
